requirements.txt
appnope==0.1.0
backcall==0.1.0
decorator==4.2.1
Django==2.0.4
django-extensions==2.0.6
djangorestframework==3.8.1
ipython==6.3.0
ipython-genutils==0.2.0
jedi==0.11.1
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2018.3
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
typing==3.6.4
wcwidth==0.1.7
